requirements.txt
asgiref==3.8.1
asttokens==2.4.1
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
Django==4.2.9
django-admin-material-dashboard==1.0.10
exceptiongroup==1.2.1
executing==2.0.1
gunicorn==21.2.0
ipykernel==6.29.4
ipython==8.25.0
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.0
parso==0.8.4
pexpect==4.9.0
Pillow
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==26.0.3
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
str2bool==1.1
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
whitenoise==6.6.0
